Got distribution build working.
This involved a lot of things. The first was including a tsconfig
(and a tslint for linting purposes), and then making the compiler
happy (and appeasing the linter). The second was getting Webpack
working. Neither were monumentous, but I still had to do some quick
web searches to recall what I was supposed to do.

I'm gonna need a boiler plate for this stuff soon...