Appeased the linter by pulling everything into separate files.
This means that yash.ts is little more than a re-export file. I tried
to edit the filename to so that we were consistent with the naming,
but GitHub doesn't seem to register file name changes if all that
changed was the casing. Which I think is problematic, but oh-well.